int to long

+++ window/tumblingWindow-potential-customers/src/main/java/com/redhat/customers/event/PotentialCustomersWereDetected.java
@@ -6,7 +6,7 @@ public class PotentialCustomersWereDetected {
 
     public PotentialCustomersWereDetected() {}
 
-    public PotentialCustomersWereDetected(String num, Integer timestamp) {
+    public PotentialCustomersWereDetected(String num, Long timestamp) {
         this.num = num;
         this.timestamp = timestamp;
     }
